requirements.txt
torch
torchvision
gradio
matplotlib
scikit-learn
tqdm
seaborn
Pillow